:wrench: more refactoring, changelog update, oauth implementation
