+++ data_api/main.py
@@ -2,12 +2,7 @@
 import argparse
 import logging
 import os
-import json
-import yaml
 import toml
-import random
-import urllib
-from markdown2 import Markdown
 from airtable import airtable
 from datetime import datetime
 
@@ -19,77 +14,64 @@ from fastapi.staticfiles import StaticFiles
 from fastapi.responses import HTMLResponse
 from starlette.middleware.cors import CORSMiddleware
 from starlette.requests import Request
-from fastapi.templating import Jinja2Templates
+from fastapi.security import OAuth2PasswordBearer, OAuth2PasswordRequestForm
 from pydantic import Json
 
-from .routers import collections
+from .routers import collections, upload
+from .deps import (
+    get_db,
+    templates,
+    logger,
+)
+from .lib_cfg import (
+    config,
+)
+import data_api.deps as deps
 
 import data_api.lib_misc as lm
-from data_api.lib_parse import (
-    convert
-)
 from data_api.models import (
-    SubmitModel,
-    ReadModel,
-    UpdateModel,
-    # ListModel,
+    ListModel,
     ListTypes,
 )
 
 
 # ################################################### SETUP AND ARGUMENT PARSING
 # ##############################################################################
-logger = logging.getLogger(__name__)
-logger.setLevel(logging.getLevelName('INFO'))
-logger.addHandler(logging.StreamHandler())
 dir_path = os.path.dirname(os.path.realpath(__file__))
 
-config = {
-    'postgresql': {
-        'dsn': os.getenv('PG_DSN', 'postgres://user:pass@localhost:5432/db'),
-        'min_size': 4,
-        'max_size': 20
-    },
-    'proxy_prefix': os.getenv('PROXY_PREFIX', ''),
-    'server': {
-        'host': os.getenv('HOST', '127.0.0.1'),
-        'port': int(os.getenv('PORT', '5000')),
-        'log_level': os.getenv('LOG_LEVEL', 'info'),
-        'timeout_keep_alive': 0,
-    },
-    'log_level': 'info',
-    'airtable': {
-        'base_id': os.getenv('AIRTABLE_BASE', ''),
-        'api_key': os.getenv('AIRTABLE_API', ''),
-    },
-    'salt': os.getenv('SALT', 'OpenJusticePirates'),
-}
 
-VERSION = 3
+VERSION = 4
 START_TIME = datetime.now(pytz.utc)
 
+tags_metadata = [
+    {
+        "name": "users",
+        "description": "Operations related to users"
+    },
+    {
+        "name": "authentication",
+        "description": "Login, logout, etc."
+    },
+    {
+        "name": "collections",
+        "description": "Operations on collections, depending authentified user"
+    },
+    {
+        "name": "upload",
+        "description": "Operations related to the upload of documents"
+    },
+]
 
-async def get_db():
-    global DB_POOL  # pylint:disable=global-statement
-    conn = await DB_POOL.acquire()
-    try:
-        yield conn
-    finally:
-        await DB_POOL.release(conn)
-
-
-def doc_hash(ecli):
-    nonce = random.randint(0, 99999999)
-    num = hex(abs(hash(F"{ecli}{nonce}{config['salt']}{datetime.now()}")))
-    return num.lstrip("0x").rstrip("L")
+oauth2_scheme = OAuth2PasswordBearer(tokenUrl=config.key('token'))
 
 # ############################################################### SERVER ROUTES
 # #############################################################################
 
 
-app = FastAPI(root_path=config['proxy_prefix'])
+app = FastAPI(root_path=config.key('proxy_prefix'), openapi_tags=tags_metadata)
 app.mount("/static", StaticFiles(directory="./static"), name="static")
 app.include_router(collections.router)
+app.include_router(upload.router)
 app.add_middleware(
     CORSMiddleware,
     allow_origins=["*"],
@@ -98,128 +80,28 @@ app.add_middleware(
     allow_headers=["*"],
 )
 
-templates = Jinja2Templates(directory="templates")
 
-
-# ############################################################### SERVER ROUTES
-# #############################################################################
 @app.on_event("startup")
 async def startup_event():
-    global DB_POOL  # pylint:disable=global-statement
     if os.getenv('NO_ASYNCPG', 'false') == 'false':
         try:
-            DB_POOL = await asyncpg.create_pool(**config['postgresql'])
+            import json
+            print(json.dumps(config.key('postgresql')))
+            cfg = config.key('postgresql')
+            deps.DB_POOL = await asyncpg.create_pool(**cfg)
         except asyncpg.InvalidPasswordError:
-            if config['log_level'] != 'debug':
+            if config.key('log_level') != 'debug':
                 logger.critical("No database found")
                 raise
             logger.warning("No Database Found !!!! But we're in debug mode, proceeding anyway")
-            DB_POOL = False
+            deps.DB_POOL = False
+# ############################################################### SERVER ROUTES
+# #############################################################################
 
 
 @app.get("/")
 def root():
     return lm.status_get(START_TIME, VERSION)
-
-
-@app.post("/create")
-async def create(query: SubmitModel, request: Request, db=Depends(get_db)):
-    """
-    Submit document endpoint
-    """
-    logger.info('Testing user key %s', query.user_key)
-    # FIXME : Fix airtable key checking
-    at = airtable.Airtable(config['airtable']['base_id'], config['airtable']['api_key'])
-    res = at.get('Test Users', filter_by_formula="FIND('%s', {Key})=1" % query.user_key)
-    ecli = f"ECLI:{query.country}:{query.court}:{query.year}:{query.identifier}"
-
-    if res and 'records' in res and len(res['records']) == 1:
-        rec = res['records'][0]['fields']
-        logger.info("User %s / %s submitting text %s", rec['Name'], rec['Email'], ecli)
-    else:
-        raise HTTPException(status_code=401, detail="bad user key")
-
-    meta = query.meta if query.meta is not None else {}
-    meta['labels'] = query.labels
-
-    docHash = doc_hash(ecli)
-
-    sql = """
-    INSERT INTO ecli_document (
-        ecli,
-        country,
-        court,
-        year,
-        identifier,
-        text,
-        meta,
-        ukey,
-        lang,
-        appeal,
-        hash
-    ) VALUES ( $1, $2, $3, $4, $5, $6, $7, $8, $9, $10, $11)
-    RETURNING id_internal;
-    """
-
-    docId = await db.fetchval(
-        sql,
-        ecli,
-        query.country,
-        query.court,
-        query.year,
-        query.identifier,
-        query.text,
-        json.dumps(meta),
-        query.user_key,
-        query.lang,
-        query.appeal,
-        docHash,
-    )
-
-    # Keep labels in database for reuse
-    for label in query.labels:
-        check = await db.fetchrow("SELECT 'one' FROM labels WHERE label = $1", label)
-        if not check:
-            logging.debug("New label : %s", label)
-            await db.execute("INSERT INTO labels (label, category) VALUES ($1, 'user_defined')", label)
-
-    # Store doclinks
-    for doc in query.doc_links:
-        sql = """
-            INSERT INTO ecli_links (
-                id_internal,
-                target_type,
-                target_identifier,
-                target_label
-            ) VALUES ($1, $2, $3, $4);
-        """
-
-        await db.execute(
-            sql,
-            docId,
-            doc.kind,
-            doc.link,
-            doc.label,
-        )
-
-    logger.debug('Wrote ecli %s ( hash %s ) to database', ecli, docHash)
-    return {'result': "ok", 'hash': docHash}
-
-
-@app.get("/read")
-def read(query: ReadModel, request: Request, db=Depends(get_db)):
-    """
-    Access document endpoint
-    """
-    raise HTTPException(status_code=501, detail="Not implemented")
-
-
-@app.get("/update")
-def update(query: UpdateModel, request: Request, db=Depends(get_db)):
-    """
-    Update document endpoint
-    """
-    raise HTTPException(status_code=501, detail="Not implemented")
 
 
 @app.get("/list")
@@ -253,98 +135,6 @@ async def getList(request: Request, db=Depends(get_db), level: ListTypes = 'coun
         raise HTTPException(status_code=404, detail="Not Found")
 
 
-@app.get("/hash/{dochash}", response_class=HTMLResponse)
-async def gohash(request: Request, dochash: str, db=Depends(get_db)):
-    sql = """
-    SELECT id_internal, ecli, text, appeal, meta->'labels' AS labels
-    FROM ecli_document
-    WHERE hash = $1
-    """
-
-    res = await db.fetchrow(sql, dochash)
-
-    if not res:
-        raise HTTPException(status_code=404, detail="Document not found")
-
-    markdowner = Markdown()
-    html_text = markdowner.convert(
-        res['text'].replace('_', '\\_')
-    )
-    html_text = convert(html_text)
-
-    sql_links = """
-    SELECT target_type, target_identifier, target_label
-    FROM ecli_links
-    WHERE id_internal = $1
-    """
-
-    res2 = await db.fetch(sql_links, res['id_internal'])
-    ecli_links = []
-    eli_links = []
-    for row in res2:
-        if row['target_type'] == 'ecli':
-            ecli_links.append({'name': row['target_label'], 'id': row['target_identifier']})
-        elif row['target_type'] == 'eli':
-            eli_links.append({'name': row['target_label'], 'link': row['target_identifier']})
-
-    return templates.TemplateResponse('share.html', {
-        'request': request,
-        'ecli': res['ecli'],
-        'text': html_text,
-        'labels': json.loads(res['labels']) if res['labels'] else [],
-        'appeal': res['appeal'],
-        'elis': eli_links,
-        'eclis': ecli_links,
-    })
-
-
-@app.get("/html/{ecli}", response_class=HTMLResponse)
-async def ecli(request: Request, ecli, db=Depends(get_db)):
-    # FIXME: add text output on request ACCEPT
-    sql = """
-    SELECT id_internal, ecli, text, appeal, meta->'labels' AS labels
-    FROM ecli_document
-    WHERE ecli = $1
-    AND status = 'public'
-    """
-
-    res = await db.fetchrow(sql, ecli)
-
-    if not res:
-        raise HTTPException(status_code=404, detail="Document not found")
-
-    markdowner = Markdown()
-    html_text = markdowner.convert(
-        res['text'].replace('_', '\\_')
-    )
-    html_text = convert(html_text)
-
-    sql_links = """
-    SELECT target_type, target_identifier, target_label
-    FROM ecli_links
-    WHERE id_internal = $1
-    """
-
-    res2 = await db.fetch(sql_links, res['id_internal'])
-    ecli_links = []
-    eli_links = []
-    for row in res2:
-        if row['target_type'] == 'ecli':
-            ecli_links.append({'name': row['target_label'], 'id': row['target_identifier']})
-        elif row['target_type'] == 'eli':
-            eli_links.append({'name': row['target_label'], 'link': row['target_identifier']})
-
-    return templates.TemplateResponse('share.html', {
-        'request': request,
-        'ecli': res['ecli'],
-        'text': html_text,
-        'labels': json.loads(res['labels']) if res['labels'] else [],
-        'appeal': res['appeal'],
-        'elis': eli_links,
-        'eclis': ecli_links,
-    })
-
-
 @app.get("/test", response_class=HTMLResponse)
 def test(request: Request, ):
     """
@@ -360,45 +150,29 @@ def test(request: Request, ):
     })
 
 
-@app.get("/labels/{begin}")
-async def labels(begin, db=Depends(get_db)):
+@app.post("/token")
+async def login(form_data: OAuth2PasswordRequestForm = Depends()):
     """
-    Return matching labels (only search from beginning of string)
+    Authentify users
     """
-    sql = """
-    SELECT label FROM labels WHERE LOWER(label) LIKE $1 || '%'
-    """
+    # FIXME: Some other auth provider then airtable would be nice
+    # FIXME: Add support for scopes (like admin, moderatore, ...)
+    # FIXME: Add password hashing support
+    at = airtable.Airtable(config.key(['airtable', 'base_id']), config.key(['airtable', 'api_key']))
+    res = at.get('Test Users', filter_by_formula="FIND('%s', {Key})=1" % form_data.password)
 
-    res = await db.fetch(sql, begin.lower())
-
-    output = []
-    for row in res:
-        output.append(row['label'])
-
-    return output
-
-
-@app.get("/c/{collection}")
-async def collection(collection, db=Depends(get_db)):
-    """
-    Collection endpoint
-    /c/[collection] lists members of a specific collection, curated by a group of moderators
-
-    Some collections have a specific meaning, and are protected:
-    /c/admin -> list everything
-    /c/review -> latests documents awaiting review
-    ...
-    """
-
-
-
+    try:
+        rec = res['records'][0]['fields']
+        assert(rec['Email'] == form_data.username)
+        logger.info("Authentified user %s", rec['Name'])
+        return {"access_token": rec['Name'], "token_type": "bearer"}
+    except (KeyError, AssertionError):
+        raise HTTPException(status_code=401, detail="Incorrect username or password")
 
 
 # ##################################################################### STARTUP
 # #############################################################################
 def main():
-    global config
-
     parser = argparse.ArgumentParser(description='Matching server process')
     parser.add_argument('--config', dest='config', help='config file', default=None)
     parser.add_argument('--debug', dest='debug', action='store_true', default=False, help='Debug mode')
@@ -410,20 +184,25 @@ def main():
     else:
         t_config = toml.load('config_default.toml')
 
-    config = {**config, **t_config}
+    config.merge(t_config)
 
     if args.debug:
         logger.setLevel(logging.getLevelName('DEBUG'))
         logger.debug('Debug activated')
-        config['log_level'] = 'debug'
-        config['server']['log_level'] = 'debug'
+        config.set('log_level', 'debug')
+        config.set(['server', 'log_level'], 'debug')
         logger.debug('Arguments: %s', args)
-        logger.debug('config: %s', yaml.dump(config, indent=2))
+        config.dump()
         # logger.debug('config: %s', toml.dumps(config))
 
+        # uvicorn.run(
+        #     "data_api.main:app",
+        #     reload=True,
+        #     **config.key('server')
+        # )
     uvicorn.run(
         app,
-        **config['server']
+        **config.key('server')
     )
