+++ data_api/routers/collections.py
@@ -1,13 +1,6 @@
 from fastapi import APIRouter
 
-tags_metadata = [
-    {
-        "name": "collections",
-        "description": "Operations on collections, depending authentified user"
-    }
-]
-
-router = APIRouter(openapi_tags=tags_metadata)
+router = APIRouter()
 
 
 @router.get("/c/", tags=["collections"])
@@ -17,6 +10,15 @@ async def read_collections():
 
 @router.get("/c/{collection}", tags=["collections"])
 async def read_collection(collection: str):
+    """
+    Collection endpoint
+    /c/[collection] lists members of a specific collection, curated by a group of moderators
+
+    Some collections have a specific meaning, and are protected:
+    /c/admin -> list everything
+    /c/review -> latests documents awaiting review
+    ...
+    """
     return {
         "collection": collection
     }
